adicionar lógica para tratar de golpes nas faixas

+++ src/service/faixa-api.service.ts
@@ -1,18 +1,27 @@
-import { Faixa } from "src/model/faixa";
+// import { RepeatedValueException } from "src/exception/exception.exception";
+import { Faixa } from "src/model/faixa.js";
 
 export class FaixaApiService {
     private faixas: Faixa[]
 
     constructor() {
-        this.faixas = JSON.parse(localStorage.getItem('faixas')) ?? []
+        const jsonFaixas = JSON.parse(localStorage.getItem('faixas')) as Faixa[] ?? []
+        this.faixas = jsonFaixas.map(jsonFaixa => new Faixa(jsonFaixa.nome))
     }
 
+    
+    addFaixa(faixa: Faixa): void {
+        if (this.faixas.find((faixaDb) => faixa.nome === faixaDb.nome)) throw new Error(`Faixa ${faixa} já existe`)
+        this.faixas.push(faixa)
+        localStorage.setItem('faixas', JSON.stringify(this.faixas))
+    }
+    
     getFaixas(): Faixa[] {
         return JSON.parse(JSON.stringify(this.faixas))
     }
 
-    addFaixa(faixa: Faixa): void {
-        this.faixas.push(faixa)
-        localStorage.setItem('faixas', JSON.stringify(this.faixas))
+    getFaixaByName(nome: string): Faixa {
+        return this.faixas.find(faixa => faixa.nome === nome)
     }
+
 }
